+++ braille/NEWS.md
@@ -1,6 +1,7 @@
 v1.14.4-SNAPSHOT
 =======
 - Update to Liblouis [3.19.0](https://github.com/liblouis/liblouis/releases/tag/v3.19.0)
+- Support for `hyphenate-character` property
 - Bugfixes
 
 v1.14.3